+++ src/Router.tsx
@@ -14,7 +14,7 @@ const SearchPage = lazy(() => import('@/pages/SearchPage'));
 const PostDetailPage = lazy(() => import('@/pages/PostDetailPage'));
 const CreatePostPage = lazy(() => import('@/pages/CreatePostPage'));
 const ChatListPage = lazy(() => import('@/domains/ChatList/Page'));
-const ChatRoomPage = lazy(() => import('@/pages/ChatRoomPage'));
+const ChatRoomPage = lazy(() => import('@/domains/ChatRoom/Page'));
 const UsageHistoryPage = lazy(() => import('@/pages/UsageHistoryPage'));
 const MyProfilePage = lazy(() => import('@/pages/MyProfilePage'));
 const LoginPage = lazy(() => import('@/pages/LoginPage'));
