+++ src/domains/ChatRoom/components/MessageList/MessageList.style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  height: 100%;
+  overflow: auto;
+
+  padding: 0 20px;
+  -webkit-overflow-scrolling: touch;
+`;
