+++ src/domains/ChatRoom/components/SystemMessage.tsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const SystemMessage = styled.div`
+  width: fit-content;
+  margin: 20px auto;
+
+  display: flex;
+  padding: 8px 16px;
+  justify-content: center;
+
+  background-color: var(--color-gray-300);
+  color: var(--color-gray-100);
+  font-size: var(--font-semi-micro);
+  font-weight: var(--weight-regular);
+
+  border-radius: 30px;
+`;
